notification-feature-featurev0.1/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosmtplib==1.1.6
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
billiard==3.6.4.0
boto3==1.26.126
botocore==1.29.126
CacheControl==0.12.11
cachetools==5.3.0
celery==5.2.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
confz==1.8.1
coverage==7.2.7
cryptography==36.0.0
Deprecated==1.2.13
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.1
filelock==3.12.0
firebase-admin==6.1.0
flower==1.0.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
humanize==4.8.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.2
loguru==0.7.0
MarkupSafe==2.1.2
msgpack==1.0.5
multidict==6.0.4
ndg-httpsclient==0.5.1
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==23.1
phonenumbers==8.12.48
pkce==1.0.3
pluggy==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.22.2
protobuf==4.23.0
pusher==3.3.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.10.4
pydantic==1.10.7
PyJWT==2.7.0
pymongo==4.3.3
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-tornasync==0.6.0.post2
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
pyuploadcare==4.0.0
PyYAML==6.0
requests==2.30.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
sib-api-v3-sdk @ git+https://github.com/sendinblue/APIv3-python-library.git@9a5807c149db7855eca346b1d7829a96481f495c
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.26.1
starlette-session==0.4.3
stream-python==5.2.1
supertokens-python==0.16.0
tldextract==3.1.0
toml==0.10.2
tornado==6.3.2
twilio==7.9.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.22.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.8
websockets==11.0.2
win32-setctime==1.1.0
wrapt==1.15.0
yarl==1.9.2
